+++ src/pages/options/options.js
@@ -8,7 +8,7 @@ import { getSetting } from 'common/interface'
 import { COLOR_MODE_CHANGE } from 'actions'
 import { getOSModeClass } from 'common/helpers'
 import { Logo } from 'components/logo/logo'
-import { Button } from 'components/button/button'
+import { Button } from 'components/button/extensions-button'
 import { FacebookIcon } from 'components/icons/icons'
 import { TwitterIcon } from 'components/icons/icons'
 import { InstagramIcon } from 'components/icons/icons'
